Restructure API v2 module
Images and documents endpoints are now defined in their respective apps
