+++ wagtail/tests/urls.py
@@ -2,16 +2,26 @@ from __future__ import absolute_import, unicode_literals
 
 from django.conf.urls import include, url
 
-from wagtail.api.v2 import urls as wagtailapi2_urls
+from wagtail.api.v2.endpoints import PagesAPIEndpoint
+from wagtail.api.v2.router import WagtailAPIRouter
 from wagtail.contrib.wagtailapi import urls as wagtailapi_urls
 from wagtail.contrib.wagtailsitemaps.views import sitemap
 from wagtail.tests.testapp import urls as testapp_urls
 from wagtail.wagtailadmin import urls as wagtailadmin_urls
 from wagtail.wagtailcore import urls as wagtail_urls
 from wagtail.wagtaildocs import urls as wagtaildocs_urls
+from wagtail.wagtaildocs.api.v2.endpoints import DocumentsAPIEndpoint
 from wagtail.wagtailimages import urls as wagtailimages_urls
+from wagtail.wagtailimages.api.v2.endpoints import ImagesAPIEndpoint
 from wagtail.wagtailimages.tests import urls as wagtailimages_test_urls
 from wagtail.wagtailsearch import urls as wagtailsearch_urls
+
+
+api_router = WagtailAPIRouter('wagtailapi_v2')
+api_router.register_endpoint('pages', PagesAPIEndpoint)
+api_router.register_endpoint('images', ImagesAPIEndpoint)
+api_router.register_endpoint('documents', DocumentsAPIEndpoint)
+
 
 urlpatterns = [
     url(r'^admin/', include(wagtailadmin_urls)),
@@ -21,7 +31,7 @@ urlpatterns = [
     url(r'^images/', include(wagtailimages_urls)),
 
     url(r'^api/', include(wagtailapi_urls)),
-    url(r'^api/', include(wagtailapi2_urls)),
+    url(r'^api/v2beta/', api_router.urls),
     url(r'^sitemap\.xml$', sitemap),
 
     url(r'^testapp/', include(testapp_urls)),
