+++ wagtail/wagtailimages/api/v2/serializers.py
@@ -1,0 +1,7 @@
+from __future__ import absolute_import, unicode_literals
+
+from wagtail.api.v2.serializers import BaseSerializer
+
+
+class ImageSerializer(BaseSerializer):
+    pass
